Ugh a lot
I fogot to do check-ins, so this is uh what it is.
Test Rig has working soundz on the 1840
She Ra has partially working - gets as far as selecting a pattern, then hangs when it goes to the blockingDelay function. My guess is TMR0 isn't working properly after going into sleep mode, but I'm not sure what's needed to revive it.