Almost
Removed Debouncer library
Fixed the playNote loop
Sooo close - PWM for arcs is noisy resulting in flubbed notes. Need to clean it up, but I think it's not possible on this chip.
Going to move up to a chip with more PWM outputs and Timers to see if we can make it work there.